+++ tests/models/test_model_input_examples.py
@@ -1,0 +1,61 @@
+import json
+import numpy as np
+import pytest
+
+from mlflow.models.signature import infer_signature
+from mlflow.models.utils import save_example
+from mlflow.types.utils import TensorsNotSupportedException
+from mlflow.utils.file_utils import TempDir
+from mlflow.utils.proto_json_utils import _dataframe_from_json
+
+
+def test_input_examples(pandas_df_with_all_types):
+    sig = infer_signature(pandas_df_with_all_types)
+    # test setting example with data frame with all supported data types
+    with TempDir() as tmp:
+        filename = save_example(tmp.path(), pandas_df_with_all_types)["artifact_path"]
+        with open(tmp.path(filename), "r") as f:
+            data = json.load(f)
+            assert set(data.keys()) == set(("columns", "data"))
+        parsed_df = _dataframe_from_json(tmp.path(filename), schema=sig.inputs)
+        assert (pandas_df_with_all_types == parsed_df).all().all()
+        # the frame read without schema should match except for the binary values
+        assert (parsed_df.drop(columns=["binary"]) == _dataframe_from_json(tmp.path(filename))
+                .drop(columns=["binary"])).all().all()
+
+    # pass the input as dictionary instead
+    with TempDir() as tmp:
+        d = {name: pandas_df_with_all_types[name].values
+             for name in pandas_df_with_all_types.columns}
+        filename = save_example(tmp.path(), d)["artifact_path"]
+        parsed_df = _dataframe_from_json(tmp.path(filename), sig.inputs)
+        assert (pandas_df_with_all_types == parsed_df).all().all()
+
+    # input passed as numpy array
+    sig = infer_signature(pandas_df_with_all_types.values)
+    with TempDir() as tmp:
+        filename = save_example(tmp.path(), pandas_df_with_all_types.values)["artifact_path"]
+        with open(tmp.path(filename), "r") as f:
+            data = json.load(f)
+            assert set(data.keys()) == set(("data",))
+        parsed_ary = _dataframe_from_json(tmp.path(filename), schema=sig.inputs).values
+        assert (pandas_df_with_all_types.values == parsed_ary).all().all()
+
+    # pass multidimensional array
+    with TempDir() as tmp:
+        example = np.array([[[1, 2, 3]]])
+        with pytest.raises(TensorsNotSupportedException):
+            filename = save_example(tmp.path(), example)["artifact_path"]
+
+    # pass multidimensional array
+    with TempDir() as tmp:
+        example = np.array([[1, 2, 3]])
+        with pytest.raises(TensorsNotSupportedException):
+            filename = save_example(tmp.path(), {"x": example, "y": example})["artifact_path"]
+
+    # pass dict with scalars
+    with TempDir() as tmp:
+        example = {"a": 1, "b": "abc"}
+        filename = save_example(tmp.path(), example)["artifact_path"]
+        parsed_df = _dataframe_from_json(tmp.path(filename))
+        assert example == parsed_df.to_dict(orient="records")[0]
